+++ auth/src/main/java/com/firebase/ui/auth/IdpResponse.java
@@ -24,7 +24,6 @@ import android.text.TextUtils;
 
 import com.firebase.ui.auth.ui.ExtraConstants;
 import com.google.firebase.auth.FacebookAuthProvider;
-import com.google.firebase.auth.GithubAuthProvider;
 import com.google.firebase.auth.GoogleAuthProvider;
 import com.google.firebase.auth.TwitterAuthProvider;
 
@@ -33,23 +32,20 @@ import com.google.firebase.auth.TwitterAuthProvider;
  */
 public class IdpResponse implements Parcelable {
     private final User mUser;
-    private final String mPhoneNumber;
     private final String mToken;
     private final String mSecret;
     private final int mErrorCode;
 
     private IdpResponse(int errorCode) {
-        this(null, null, null, null, errorCode);
+        this(null, null, null, errorCode);
     }
 
     private IdpResponse(
             User user,
-            String phoneNumber,
             String token,
             String secret,
             int errorCode) {
         mUser = user;
-        mPhoneNumber = phoneNumber;
         mToken = token;
         mSecret = secret;
         mErrorCode = errorCode;
@@ -102,7 +98,7 @@ public class IdpResponse implements Parcelable {
      */
     @Nullable
     public String getPhoneNumber() {
-        return mPhoneNumber;
+        return mUser.getPhoneNumber();
     }
 
     /**
@@ -136,7 +132,6 @@ public class IdpResponse implements Parcelable {
     @Override
     public void writeToParcel(Parcel dest, int flags) {
         dest.writeParcelable(mUser, flags);
-        dest.writeString(mPhoneNumber);
         dest.writeString(mToken);
         dest.writeString(mSecret);
         dest.writeInt(mErrorCode);
@@ -147,7 +142,6 @@ public class IdpResponse implements Parcelable {
         public IdpResponse createFromParcel(Parcel in) {
             return new IdpResponse(
                     in.<User>readParcelable(User.class.getClassLoader()),
-                    in.readString(),
                     in.readString(),
                     in.readString(),
                     in.readInt()
@@ -163,17 +157,11 @@ public class IdpResponse implements Parcelable {
     @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
     public static class Builder {
         private User mUser;
-        private String mPhoneNumber;
         private String mToken;
         private String mSecret;
 
         public Builder(@NonNull User user) {
             mUser = user;
-        }
-
-        public Builder setPhoneNumber(String phoneNumber) {
-            mPhoneNumber = phoneNumber;
-            return this;
         }
 
         public Builder setToken(String token) {
@@ -190,8 +178,7 @@ public class IdpResponse implements Parcelable {
             String providerId = mUser.getProviderId();
             if ((providerId.equalsIgnoreCase(GoogleAuthProvider.PROVIDER_ID)
                     || providerId.equalsIgnoreCase(FacebookAuthProvider.PROVIDER_ID)
-                    || providerId.equalsIgnoreCase(TwitterAuthProvider.PROVIDER_ID)
-                    || providerId.equalsIgnoreCase(GithubAuthProvider.PROVIDER_ID))
+                    || providerId.equalsIgnoreCase(TwitterAuthProvider.PROVIDER_ID))
                     && TextUtils.isEmpty(mToken)) {
                 throw new IllegalStateException(
                         "Token cannot be null when using a non-email provider.");
@@ -202,7 +189,7 @@ public class IdpResponse implements Parcelable {
                         "Secret cannot be null when using the Twitter provider.");
             }
 
-            return new IdpResponse(mUser, mPhoneNumber, mToken, mSecret, ResultCodes.OK);
+            return new IdpResponse(mUser, mToken, mSecret, ResultCodes.OK);
         }
     }
 }
