+++ auth/src/main/java/com/firebase/ui/auth/ui/phone/PhoneVerificationActivity.java
@@ -289,8 +289,9 @@ public class PhoneVerificationActivity extends AppCompatBase {
 
     private void finish(FirebaseUser user) {
         IdpResponse response = new IdpResponse.Builder(
-                new User.Builder(PhoneAuthProvider.PROVIDER_ID, null).build())
-                .setPhoneNumber(user.getPhoneNumber())
+                new User.Builder(PhoneAuthProvider.PROVIDER_ID, null)
+                        .setPhoneNumber(user.getPhoneNumber())
+                        .build())
                 .build();
         setResult(ResultCodes.OK, response.toIntent());
         finish();
